refactor: extract shared utilities to remove duplication
After /simplify review, applied the following clean-ups:

- Extract derivePlatformOrigin from server.ts into utils/corsOrigins.ts;
  test now imports the real function instead of replicating it.
- Extract promptInput/promptPassword into scripts/lib/prompts.ts and reuse
  in both bootstrap-admin.ts (-65 lines) and create-platform-admin.ts (-44).
- Extract useSidebarCollapse hook for the localStorage-backed collapse state.
  Used in both Layout.tsx and PlatformLayout.tsx; the hook also skips the
  redundant initial-mount localStorage write that both inlined versions did.
- Consolidate validateInput's overlapping password length conditions in
  create-platform-admin.ts into a single passwordRequired/passwordProvided
  guard.
- Drop the {...rest, email: tempEmail} spread+overwrite anti-pattern; the
  interactive prompt now takes an email arg directly.

Net: -67 lines, no behavior change. Backend build clean, 11/11 cors tests
still pass, frontend build clean.
