+++ backend/scripts/bootstrap-admin.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from '@prisma/client';
 import bcrypt from 'bcrypt';
-import * as readline from 'readline';
+import { promptInput, promptPassword } from './lib/prompts';
 
 const prisma = new PrismaClient();
 
@@ -23,72 +23,6 @@ interface AdminInput {
  * 2. Environment variables: Set BOOTSTRAP_ADMIN_EMAIL, BOOTSTRAP_ADMIN_PASSWORD, etc.
  */
 
-async function promptForInput(question: string): Promise<string> {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-
-  return new Promise((resolve) => {
-    rl.question(question, (answer) => {
-      rl.close();
-      resolve(answer.trim());
-    });
-  });
-}
-
-/**
- * Prompt for password with masked input (shows * for each character)
- * Uses Node.js readline in raw mode for secure input
- */
-async function promptForPassword(question: string): Promise<string> {
-  return new Promise((resolve) => {
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-
-    let password = '';
-    process.stdout.write(question);
-
-    // Enable raw mode to capture individual keystrokes
-    if (process.stdin.isTTY) {
-      process.stdin.setRawMode(true);
-    }
-
-    process.stdin.on('data', (char) => {
-      const str = char.toString('utf-8');
-
-      // Handle different key inputs
-      if (str === '\n' || str === '\r' || str === '\u0004') {
-        // Enter or Ctrl+D - finish input
-        if (process.stdin.isTTY) {
-          process.stdin.setRawMode(false);
-        }
-        rl.close();
-        process.stdout.write('\n');
-        resolve(password);
-      } else if (str === '\u0003') {
-        // Ctrl+C - cancel
-        if (process.stdin.isTTY) {
-          process.stdin.setRawMode(false);
-        }
-        process.stdout.write('\n');
-        process.exit(0);
-      } else if (str === '\u007f' || str === '\b') {
-        // Backspace - remove last character
-        if (password.length > 0) {
-          password = password.slice(0, -1);
-          process.stdout.write('\b \b');
-        }
-      } else if (str >= ' ' && str <= '~') {
-        // Printable character - add to password
-        password += str;
-        process.stdout.write('*');
-      }
-    });
-  });
-}
 
 async function getAdminInputFromEnv(): Promise<AdminInput | null> {
   const email = process.env.BOOTSTRAP_ADMIN_EMAIL;
@@ -108,11 +42,11 @@ async function getAdminInputInteractive(): Promise<AdminInput> {
   console.log('\n📝 Please provide admin user details:');
   console.log('━'.repeat(50));
 
-  const email = await promptForInput('Email address: ');
-  const password = await promptForPassword('Password (min 8 characters): ');
-  const firstName = await promptForInput('First name (default: Admin): ') || 'Admin';
-  const lastName = await promptForInput('Last name (default: User): ') || 'User';
-  const phoneNumber = await promptForInput('Phone number (default: +1234567890): ') || '+1234567890';
+  const email = await promptInput('Email address: ');
+  const password = await promptPassword('Password (min 8 characters): ');
+  const firstName = await promptInput('First name (default: Admin): ') || 'Admin';
+  const lastName = await promptInput('Last name (default: User): ') || 'User';
+  const phoneNumber = await promptInput('Phone number (default: +1234567890): ') || '+1234567890';
 
   return { email, password, firstName, lastName, phoneNumber };
 }
