+++ backend/src/server.ts
@@ -40,6 +40,7 @@ import logger from './utils/logger';
 import { validateEnvironment } from './config/validateEnv';
 import { setSocketInstance } from './utils/socketInstance';
 import { validateGLAccountCodes } from './config/glAccounts';
+import { derivePlatformOrigin } from './utils/corsOrigins';
 
 // Routes
 import authRoutes from './routes/authRoutes';
@@ -144,21 +145,6 @@ app.use('/api/public', (_req, res, next) => {
 // SECURITY: Trust only the first proxy (nginx). Using `true` trusts ALL proxies,
 // allowing attackers to spoof X-Forwarded-For and bypass rate limiting.
 app.set("trust proxy", 1);
-
-// Auto-derive platform subdomain from FRONTEND_URL so a missing PLATFORM_URL
-// env doesn't silently break platform.<apex> logins. Explicit PLATFORM_URL
-// still wins when set.
-function derivePlatformOrigin(frontendUrl: string): string | null {
-  try {
-    const url = new URL(frontendUrl);
-    if (!url.hostname || url.hostname === 'localhost' || url.hostname.startsWith('platform.')) {
-      return null;
-    }
-    return `${url.protocol}//platform.${url.hostname}`;
-  } catch {
-    return null;
-  }
-}
 
 const frontendUrl = process.env.FRONTEND_URL || 'http://localhost:5173';
 const platformUrl = process.env.PLATFORM_URL || derivePlatformOrigin(frontendUrl);
